requirements.txt
python-dotenv
pandas
numpy
torch
pytorch_lightning
tensorboard
fastapi
celery
redis
uvicorn